# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise HRMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{organizationId}/recruitment-settings:
    get:
      operationId: getRecruitmentSettings
      tags: [recruitment]
      summary: Get an organization's Recruitment settings
      description: >-
        Returns the organization's saved Recruitment settings, or safe
        defaults if nothing has been saved yet — no row is created on read.
        Gated by the recruitment module and recruitment_settings.read.
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recruitment settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecruitmentSettings"
    patch:
      operationId: updateRecruitmentSettings
      tags: [recruitment]
      summary: Update an organization's Recruitment settings
      description: >-
        Merges the given fields into the organization's existing settings
        (or defaults) and validates the merged result before saving. Gated
        by the recruitment module and recruitment_settings.manage.
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecruitmentSettingsInput"
      responses:
        "200":
          description: Updated Recruitment settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecruitmentSettings"
        "400":
          description: Invalid settings, or defaultWorkflowId does not belong to this organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RecruitmentSettings:
      type: object
      description: >-
        One row per organization. `updatedAt` is null when nothing has ever
        been saved — the response is still safe defaults, not an error.
      required:
        - organizationId
        - enabled
        - internalRecruitmentEnabled
        - externalRecruitmentEnabled
        - requireCandidateAccount
        - defaultWorkflowId
        - candidateDataRetentionMonths
        - reapplicationWaitingDays
        - duplicateCandidatePolicy
        - defaultOfferExpiryDays
        - defaultDocumentRequirements
        - applicationLimitPerCandidate
        - updatedAt
      properties:
        organizationId:
          type: integer
        enabled:
          type: boolean
        internalRecruitmentEnabled:
          type: boolean
        externalRecruitmentEnabled:
          type: boolean
        requireCandidateAccount:
          type: boolean
        defaultWorkflowId:
          type: ["integer", "null"]
        candidateDataRetentionMonths:
          type: ["integer", "null"]
        reapplicationWaitingDays:
          type: integer
        duplicateCandidatePolicy:
          type: string
          enum: [allow, flag, block]
        defaultOfferExpiryDays:
          type: ["integer", "null"]
        defaultDocumentRequirements:
          type: array
          items:
            type: string
          description: Master Data categoryCode strings, free text — not validated against a domain's item list.
        applicationLimitPerCandidate:
          type: ["integer", "null"]
        updatedAt:
          type: ["string", "null"]
          format: date-time
    UpdateRecruitmentSettingsInput:
      type: object
      description: All fields optional — merged into the organization's existing settings (or defaults) before validation.
      properties:
        enabled:
          type: boolean
        internalRecruitmentEnabled:
          type: boolean
        externalRecruitmentEnabled:
          type: boolean
        requireCandidateAccount:
          type: boolean
        defaultWorkflowId:
          type: ["integer", "null"]
        candidateDataRetentionMonths:
          type: ["integer", "null"]
        reapplicationWaitingDays:
          type: integer
          minimum: 0
        duplicateCandidatePolicy:
          type: string
          enum: [allow, flag, block]
        defaultOfferExpiryDays:
          type: integer
          minimum: 1
        defaultDocumentRequirements:
          type: array
          items:
            type: string
        applicationLimitPerCandidate:
          type: integer
          minimum: 1
...